wip: StdCircuit
